improved coverage structure
